feat: accelerated LinEquSolve with cuSolver LU factor
first version, just to show a POC that it's worth pursuing this route.
included a makefile for compilation without leaving the directory dirty.
to show that the solution is correct, i calculate the difference between the two solution,
and find the maximum absolute distance between them (zero).
the time savings are impressive (~80x), but the code has to be reviewed to be more readable
and less shitty. sorry for the single big commit, i'll break it up in pieces next time.