refactor: split into different files
this was way harder than i thought! classic build systems getting in the way of building...
i think this makes a bit more sense now. the directory structure is the next thing that should
be reviewed, but it can wait. and the coolest thing would be to have a smart makefile that
finds the most recent version of cuda supported by current hardware... :)